start offset tracking during writing

+++ lead.h
@@ -33,6 +33,6 @@ struct lead {
 int lead_init(struct lead *lead, int fd);
 void lead_destroy(struct lead *lead);
 void lead_dump(const struct lead *lead, FILE *f);
-int lead_write(const struct lead *lead, int fd, off_t ofs);
+off_t lead_write(const struct lead *lead, int fd, off_t ofs);
 
 #endif
